requirements.txt
pyyaml>=6.0
speedtest-cli>=2.1.3
gspread>=6.0.0
google-auth>=2.0.0
flask>=3.0.0
croniter>=2.0.0
apscheduler>=3.10.0
